requirements.txt
langchain==0.3.25
langchain-groq==0.2.1
langchain-community==0.3.23
langgraph==0.4.1
langsmith==0.3.42
pymupdf==1.25.5
sentence-transformers==2.7.0
scikit-learn==1.6.1
pydantic==2.11.4
python-dotenv==1.1.0
rich==13.9.4
